final 1 commit

+++ src/app/video-card-delete/video-card-delete.component.ts
@@ -11,22 +11,33 @@ import { UserService } from '../user.service';
 export class VideoCardDeleteComponent {
   @Input()
   video!: VideoDto;
+  img!: string;
 
 
   myVideos: Array<VideoDto> = []
 
-  constructor(private videoService: VideoService, private userService: UserService) {}
+  constructor(private videoService: VideoService, private userService: UserService) {
+    this.userService.getUserById(this.userService.userIdnew).subscribe(data1 => {
+      this.img = data1.image
+    })
+  }
 
   deleteVideo(id: string) {
-    alert("Xác nhận xóa video")
-    this.videoService.deleteVideo(id).subscribe(data => {
-      console.log("xác nhận xóa thành công")
-    })
-
-    this.videoService.getVideosByUserId(this.userService.userIdnew).subscribe(response => {
-      console.log(this.myVideos)
-      this.myVideos = response
-    })
+    const isConfirmed = confirm("Bạn muốn xóa video?")
+    if(isConfirmed) {
+      this.videoService.deleteVideo(id).subscribe(data => {
+        console.log("xác nhận xóa thành công")
+        alert("Xác nhận xóa thành công!")
+      })
+  
+      this.videoService.getVideosByUserId(this.userService.userIdnew).subscribe(response => {
+        console.log(this.myVideos)
+        this.myVideos = response
+      })
+    }
+    else {
+      alert("Không xóa video!")
+    }
   }
 
 }
